Move people-autocomplete-wrapper into separate folder
