+++ src/ggrc-client/js/components/people-autocomplete/people-autocomplete-wrapper/people-autocomplete-wrapper.js
@@ -3,10 +3,10 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import './people-autocomplete-results/people-autocomplete-results';
+import '../people-autocomplete-results/people-autocomplete-results';
 
-import baseAutocompleteWrapper from './../custom-autocomplete/autocomplete-wrapper';
-import PersonModel from '../../models/business-models/person';
+import baseAutocompleteWrapper from '../../custom-autocomplete/autocomplete-wrapper';
+import PersonModel from '../../../models/business-models/person';
 
 let viewModel = baseAutocompleteWrapper.extend({
   currentValue: null,
