+++ src/ggrc-client/js/components/rich-text/people-mention-controller/people-mention-controller.js
@@ -3,7 +3,7 @@
   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import '../../people-autocomplete/people-autocomplete-wrapper';
+import '../../people-autocomplete/people-autocomplete-wrapper/people-autocomplete-wrapper';
 
 import template from './people-mention-controller.stache';
 import {KEY_MAP} from '../../custom-autocomplete/autocomplete-input';
